+++ src/pyramid_bimt/tests/test_user_views.py
@@ -522,7 +522,10 @@ class TestUserAdd(unittest.TestCase):
         'valid_to': date(2014, 2, 1),
         'last_payment': date(2014, 1, 1),
         'groups': [1, ],
-        'properties': [{u'key': u'foo', u'value': u'bar'}, ],
+        'properties': [
+            {u'key': u'foo', u'value': u'bar'},
+            {u'key': u'affiliate_id'},
+        ],
     }
 
     def setUp(self):
@@ -630,7 +633,9 @@ class TestUserEdit(unittest.TestCase):
             'valid_to': date(2014, 2, 1),
             'last_payment': date(2014, 1, 1),
             'groups': ['1', ],
-            'properties': [{'key': u'foo', 'value': u'bar'}, ],
+            'properties': [
+                {'key': u'foo', 'value': u'bar'},
+            ],
         })
 
     def test_view_csrf_token(self):
@@ -650,6 +655,7 @@ class TestUserEdit(unittest.TestCase):
         'properties': [
             {u'key': u'foo', u'value': u'bar'},   # existing property
             {u'key': u'baz', u'value': u'bam'},   # new property
+            {u'key': u'empty'},   # new property
         ],
     }
 
@@ -674,6 +680,7 @@ class TestUserEdit(unittest.TestCase):
         self.assertEqual(user.groups, [Group.by_id(1), Group.by_id(3)])  # enabled user stays enabled  # noqa
         self.assertEqual(user.get_property('foo'), 'bar')
         self.assertEqual(user.get_property('baz'), 'bam')
+        self.assertEqual(user.get_property('empty'), None)
         with self.assertRaises(KeyError):
             user.get_property('bimt')  # removed property